Add skeleton of title page
At this step, I defined the structure of the BML page with all the elements
that I expect to appear in this page, and "semantic" classes that describe
the types of elements. I expect to use these semantic classes, not for styling
but as a basis for template matching in XSLT transformations to generate
separate versions of the book for HTML browsers on PC and for print.