Initial layout for the title page
Only the title of the book is visible. I plan to make the title page
interactive, with an animation to display the names of authors and
the license after some time.

There are 4 colored areas in the corners, which complete the 6 colors
of the Color opponency theory: red vs green, blue vs yellow, and the
title in the center is set in black vs white.

These 4 colors are also the colors of the 4 buttons on the remote.
I plan to add an Easter Egg game similar to the iconic Simon Game:
  http://en.wikipedia.org/wiki/Simon_%28game%29
When one of the colored buttons on the remote is pressed, the corresponding
area is highlighted and a sound is played.

There is no time limit, and free play is encouraged: in case of error,
the same sequence is repeated. The sequence is considered valid even if
extra notes are included before, after, or in between. The machine waits
for several seconds of silence before repeating the melody or playing
the extended melody.